Fix wrong number of pages when copying

+++ swiper/swiper.py
@@ -40,7 +40,8 @@ def print_file(user, file):
     copies = 1
     fout = os.path.dirname(file) + '/' + uuid.uuid4().hex + '.pdf'
     gsopts = [
-        '-dNOPAUSE', '-dBATCH', '-sDEVICE=pdfwrite', '-sPAPERSIZE=a4', '-sOutputFile=' + fout
+        '-dNOPAUSE', '-dBATCH', '-sDEVICE=pdfwrite', '-sPAPERSIZE=a4',
+        '-sOutputFile=' + fout
     ]
     cmd = ['/sbin/runuser', '-u', user, '--', '/usr/bin/gs', *gsopts, file]
     p = subprocess.Popen(cmd)
@@ -49,33 +50,28 @@ def print_file(user, file):
     ps = conn.getPrinters()
     ps = list(ps.keys())
     lpopts = ['-o', 'media=A4', '-t', os.path.basename(file)]
-    lpopts += ['-o', 'page-left=0 page-right=0 page-top=0 page-bottom=0 scaling=94']
+    lpopts += [
+        '-o', 'page-left=0 page-right=0 page-top=0 page-bottom=0 scaling=94'
+    ]
     m = re.search('copias=([0-9]+)', file)
     if m:
-        copies = int(m.groups(1)[0])
+        lpopts += ['-n', m.groups(1)[0]]
     if 'dupla-face' in file:
         lpopts += ['-o', 'sides=two-sided-long-edge']
     m = [p for p in ps if p in file]
     if m:
         lpopts += ['-d', max(m)]
-    while copies > 0:
-        copies -= 1
-        cmd = ['/sbin/runuser', '-u', user, '--', '/usr/bin/lp', *lpopts, fout]
-        print(cmd)
-        p = subprocess.Popen(cmd)
-        p.wait()
+    cmd = ['/sbin/runuser', '-u', user, '--', '/usr/bin/lp', *lpopts, fout]
+    p = subprocess.Popen(cmd)
+    p.wait()
     remove(fout)
 
 
 def remove(file):
-    try:
-        os.remove(file)
-    except Exception as e:
-        print(e)
-    try:
+    if os.path.isdir(file):
         shutil.rmtree(file, ignore_errors=True, onerror=None)
-    except Exception as e:
-        print(e)
+    else:
+        os.remove(file)
 
 
 while True:
